requirements.txt
pyinstaller
pyautogui
pystray
Pillow